generator: generate overlapping symbols
Generate symbols that overlap the known symbol:

    ---   0b111   <known symbol>.7
    --    0b110   <known symbol>.6
    -     0b100   <known symbol>.4
     --   0b011   <known symbol>.3
     -    0b010   <known symbol>
      -   0b001   <known symbol>.1

This test the ability for dladdr() to locate the best symbol:
an address in <known symbol> also match <known symbol>.7, .6,
and .3. But one would expect dladdr() to return <known symbol>,
eg. the smallest one matching.

As the size requirement is tripled, the maximum size has to
be reduced in Makefile so to be able to link below 2GiBytes.